Implement DOO counter in UE4
RefereeBlueprint is a actor component that attaches to Referee and interacts
with the gameworld on its behalf. Spilne cones blueprint has been modified,
so that it reports to a Referee class whenever a cone was hit.

WARNING: Probably bad stuff will happen if there are more than one referees on the map or
if a Referee (C++ class) without an attached RefereeBlueprint exist